Big overhaul, transitioned game from BasicGame to StateBasedGame
After many failed iterations of the flashlights I'm going to go for a
simpler method for the guards detecting the player. Game is now a
StateBasedGame to allow for a menu screen, multiple levels, and a
victory/defeat screen. Player movement also smoothed out a bit. Got new
artwork and animations for the player and guards.